Redundancy checker

+++ get_catalog.py
@@ -11,7 +11,7 @@ def argument_parser():
     result = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     # path to the config file with parameters and information about the run
     result.add_argument('-c', dest='cutouts', type=str, default="catalogs.txt") 
-    result.add_argument('-d', dest='directory', type=str, default="catalogs") 
+    result.add_argument('-d', dest='directory', type=str, default="catalogs.nosync") 
     result.add_argument('-l', dest='layer', type=str, default="ls-dr9")
 
     return result
@@ -35,6 +35,8 @@ if __name__ == '__main__':
 
     url = 'https://www.legacysurvey.org/viewer/{}/cat.fits?ralo={}&rahi={}&declo={}&dechi={}'
 
+    socket.setdefaulttimeout(15)
+
     # loop through each cutout and download the data
     for cutout in cutout_list:
         ra, dec, size = cutout
@@ -43,15 +45,16 @@ if __name__ == '__main__':
 
         outname = '/'+str(ra)+'_'+str(dec)+'.fits'
 
-        socket.setdefaulttimeout(15)
+        if not os.path.isfile(outputs+outname):
+            try:
+                urlib.urlretrieve(url.format(args.layer, ra_min, ra_max, dec_min, dec_max),outputs+outname)
+            except Exception as e:
+                print('Catalog at [', ra, dec, '] timed out. :(')
+                continue
 
-        try:
-            urlib.urlretrieve(url.format(args.layer, ra_min, ra_max, dec_min, dec_max),outputs+outname)
-        except Exception as e:
-            print('Catalog at [', ra, dec, '] timed out. :(')
-            continue
-
-        print('Catalog at [', ra, dec, '] has been downloaded.')
+            print('Catalog at [', ra, dec, '] has been downloaded.')
+        else:
+            print('Catalog at [', ra, dec, '] was already downloaded.')
     
     print('---------')
     print('Done.')
